+++ rafted_test.go
@@ -12,7 +12,7 @@ import (
 const (
     HeartbeatTimeout                        = time.Millisecond * 200
     ElectionTimeout                         = time.Millisecond * 50
-    ElectionTimeoutThresholdPersent float32 = 0.8
+    ElectionTimeoutThresholdPersent float64 = 0.8
     PersistErrorNotifyTimeout               = time.Millisecond * 100
     MaxAppendEntriesSize            uint64  = 10
     MaxSnapshotChunkSize            uint64  = 1000
